+++ src/Header.jsx
@@ -1,5 +1,4 @@
 import { useState } from "react"
-import NavLink from "./NavLink"
 import HamburgerIcon from "./icons/bars-solid.svg?react"
 import CloseIcon from "./icons/xmark-solid.svg?react"
 
@@ -7,6 +6,14 @@ function Header() {
   const [isNavOpen, setIsNavOpen] = useState(false)
   function toggleNav() {
     setIsNavOpen((prevNav) => !prevNav)
+  }
+
+  function NavLink({ text, link }) {
+    return (
+      <li className="md:px-2 md:py-0 md:last:pr-0 p-5 md:text-xl text-2xl text-purple-400 hover:text-neutral-50 transition duration-300">
+        <a href={link}>{text}</a>
+      </li>
+    )
   }
 
   return (
